// frontend/Angular/src/app/generated/api/model/pizzaModel.ts
/**
 * Pizza Ordering API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PizzaModel { 
    id?: number;
    name?: string;
    dateCreated?: string;
    disabled?: boolean;
}