// frontend/Angular/src/app/generated/api/model/getAllPizzasQuery.ts
/**
 * Pizza Ordering API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface GetAllPizzasQuery { 
    name?: string;
    disabled?: boolean;
    pageNumber?: number;
    pageSize?: number;
}